- Split LocationManager for two classes - LocationStatusBroadcastReceiver and OnLocationStatusProviderListener added to check the provider status
